requirements.txt
python-dotenv
elasticsearch
torch
sentence_transformers
fastapi
umap-learn
